cisco_meraki: improve ttl cache settings
Use the time span formspec for improved setting of cache ttl value. Fix
ttl values as they should be in seconds not minutes. Added a test that
validates that the output is cached between runs and a help text to
inform the user why they may want to override the TTL value.

+++ cmk/plugins/cisco_meraki/lib/cache.py
@@ -17,7 +17,7 @@ _HASH_NAMESPACE: Final = uuid.UUID("be8bc5b5-a401-4acc-b8db-a875cca98c79")
 _KWARG_MARK: Final = object()
 
 
-def cache_ttl[**P, R](store: Storage, *, ttl: int) -> Callable[[Callable[P, R]], Callable[P, R]]:
+def cache_ttl[**P, R](store: Storage, *, ttl: float) -> Callable[[Callable[P, R]], Callable[P, R]]:
     if ttl < 0:
         raise ValueError("Time to live value must be a positive integer.")
